fix for 64-bits platform: unsigned long instead of unsigned int when casting ptr into int
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@352 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-qt/src/DataSetDialog.cpp
@@ -107,7 +107,7 @@ namespace {
     while( it->hasNext() ) {
       pair<string,DataType> p;
       p = it->next();
-      cout << p.first << " : " << int( p.second.value ) << endl;
+      cout << p.first << " : " << (unsigned long)p.second.value << endl;
     } delete it;
   }
